shift sqlalchemy connection into tables. refactor test to use sqlalchemy for truncating tables
