+++ docker-compose.development.yml
@@ -31,7 +31,7 @@ services:
       - YACS_API_HOST=http://yacs_api:5000
 
   yacs_api:
-    command: /bin/bash -c "python database_session.py && PYTHONPATH=. alembic upgrade head && python app.py"
+    command: /bin/bash -c "python tables/database_session.py && PYTHONPATH=. alembic upgrade head && python app.py"
     ports:
       - 5000:5000
     volumes:
