+++ scripts/test.sh
@@ -1,3 +1,3 @@
 #!/bin/bash
 
-pytest --cov=. --cov-config=.coveragec -s --ignore=tests/test_data.py --ignore=tests/integration tests/
+pytest --cov=. --cov-config=.coveragec -s --ignore=tests/integration tests/
