+++ src/api/tables/__init__.py
@@ -10,3 +10,6 @@ from .student_course_selection import StudentCourseSelection
 from .user_account import UserAccount
 from .user_event import UserEvent
 from .user_session import UserSession
+
+from .database import Base
+from .database_session import SessionLocal
