+++ src/api/tables/admin_settings.py
@@ -1,7 +1,7 @@
 from sqlalchemy import Column, PrimaryKeyConstraint
 from sqlalchemy.dialects.postgresql import VARCHAR
 
-from database import Base
+from .database import Base
 
 class AdminSettings(Base):
     __tablename__ = 'admin_settings'
