+++ src/api/tables/course_prerequisite.py
@@ -1,7 +1,7 @@
 from sqlalchemy import Column, PrimaryKeyConstraint
 from sqlalchemy.dialects.postgresql import INTEGER, VARCHAR
 
-from database import Base
+from .database import Base
 
 class CoursePrerequisite(Base):
     __tablename__ = "course_prerequisite"
